+++ packages/toggle/__tests__/playwright/playwright.spec.js
@@ -1,0 +1,14 @@
+const { test, expect } = require('@playwright/test');
+import AxeBuilder from '@axe-core/playwright';
+
+test.beforeEach(async ({ page }) => {
+	await page.goto('/');
+});
+
+test.describe('Toggle > Axe', { tag: '@reduced'}, () => {
+	test('Should not have any automatically detectable accessibility issues', async ({ page }) => {	
+		const accessibilityScanResults = await new AxeBuilder({ page }).analyze(); 
+		expect(accessibilityScanResults.violations).toEqual([]);
+	});
+});
+
