+++ packages/toggle/example/src/index.html
@@ -1,137 +1,162 @@
 <!doctype html>
-<html>
-  <head>
+<html lang="en">
+
+<head>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>StormID</title>
     <style>
         * {
-            margin:0;
-            padding:0;
-            box-sizing:border-box;
+            margin: 0;
+            padding: 0;
+            box-sizing: border-box;
             font-family: sans-serif;
         }
+
         html,
         body {
-            height:100%;
+            height: 100%;
         }
+
         hr {
-            height:1px;
-            width:100%;
-            display:block;
-            background:#ddd;
-            margin:24px 0;
-            border:0 none;
+            height: 1px;
+            width: 100%;
+            display: block;
+            background: #ddd;
+            margin: 24px 0;
+            border: 0 none;
         }
+
         h1 {
             font: 1.25rem/1.2 sans-serif;
-            margin-bottom:1.5rem;
+            margin-bottom: 1.5rem;
         }
+
         h2 {
             font: 1.1rem/1.2 sans-serif;
-            margin-bottom:12px;
-            padding-top:12px;
+            margin-bottom: 12px;
+            padding-top: 12px;
         }
+
         h3 {
             font: 1rem/1.2 sans-serif;
-            margin-bottom:12px;
-            padding-top:12px;
+            margin-bottom: 12px;
+            padding-top: 12px;
         }
+
         p {
-            margin-bottom:1.5rem;
+            margin-bottom: 1.5rem;
         }
+
         header {
-            width:100%;
-            padding:20px;
-            margin-bottom:40px;
+            width: 100%;
+            padding: 20px;
+            margin-bottom: 40px;
         }
+
         .container {
-            clear:both;
-            overflow:hidden;
-            margin:0 auto;
-            width:80%;
-            max-width:1200px;
+            clear: both;
+            overflow: hidden;
+            margin: 0 auto;
+            width: 80%;
+            max-width: 1200px;
         }
+
         .row {
-            margin-left:-30px;
-            overflow:hidden
+            margin-left: -30px;
+            overflow: hidden
         }
+
         section {
-            margin-bottom:30px;
+            margin-bottom: 30px;
         }
+
         button {
-            background-color:transparent;
-            padding:0;
-            border:0 none;
-            cursor:pointer;
-            font-size:.85rem;
+            background-color: transparent;
+            padding: 0;
+            border: 0 none;
+            cursor: pointer;
+            font-size: .85rem;
         }
+
         button svg {
-            fill:blue;
+            fill: blue;
             position: relative;
-            top:.4rem;
+            top: .4rem;
         }
+
         .eg__local a {
-            margin-bottom:1rem;
-            color:blue;
+            margin-bottom: 1rem;
+            color: blue;
         }
+
         .eg__local-target {
-            max-height:0;
-            overflow:hidden;
+            max-height: 0;
+            overflow: hidden;
             transition: maxHeight 100ms ease;
         }
+
         .eg__local.is--active .eg__local-target {
-            max-height:1000px;
+            max-height: 1000px;
         }
-        .eg__local.is--active .toggle__local {
-            /* visibility:hidden; */
-        }
+
         .eg__nav {
             position: fixed;
-            top:0;
-            right:0;
-            bottom:0;
-            transform:translateX(100%);
+            top: 0;
+            right: 0;
+            bottom: 0;
+            transform: translateX(100%);
             transition: transform 120ms;
-            background:#191919;
-            width:50%;
-            padding:50px 5%;
+            background: #191919;
+            width: 50%;
+            padding: 50px 5%;
             visibility: hidden;
         }
+
         .eg__nav-item {
             display: block;
-            color:#fff;;
+            color: #fff;
+            ;
         }
+
         .on--nav .eg__nav {
             transform: translateX(0);
             visibility: visible;
         }
+
         .eg__nav-close {
             position: fixed;
-            top:20px;
-            right:20px;
+            top: 20px;
+            right: 20px;
         }
+
         .eg__nav-close svg {
             fill: #fff;
         }
     </style>
-  </head>
-  <body>
-      
-      <button class="js-toggle__btn" aria-label="Toggle off-canvas navigation">
-        <svg fill="#191919" height="24" viewBox="0 0 24 24" width="24" xmlns="http://www.w3.org/2000/svg">
-            <path d="M0 0h24v24H0z" fill="none"/>
-            <path d="M3 18h18v-2H3v2zm0-5h18v-2H3v2zm0-7v2h18V6H3z"/>
-        </svg> Menu
-    </button>
-    <nav id="nav" class="eg__nav js-toggle" data-toggle="js-toggle__btn">
-        <a class="eg__nav-item" href="#">Home</a>
-        <a class="eg__nav-item" href="#">About</a>
-        <a class="eg__nav-item" href="#">Contact</a>
-        <button class="js-toggle__btn eg__nav-close" aria-label="Toggle off-canvas navigation">
+</head>
+
+<body>
+    <main class="container">
+        <h1>Toggle example</h1>
+        <button class="js-toggle__btn" aria-label="Toggle off-canvas navigation">
             <svg fill="#191919" height="24" viewBox="0 0 24 24" width="24" xmlns="http://www.w3.org/2000/svg">
-                <path d="M19 6.41L17.59 5 12 10.59 6.41 5 5 6.41 10.59 12 5 17.59 6.41 19 12 13.41 17.59 19 19 17.59 13.41 12z"/>
-                <path d="M0 0h24v24H0z" fill="none"/>
-            </svg>
+                <path d="M0 0h24v24H0z" fill="none" />
+                <path d="M3 18h18v-2H3v2zm0-5h18v-2H3v2zm0-7v2h18V6H3z" />
+            </svg> Menu
         </button>
-    </nav>
-  </body>
+        <nav id="nav" class="eg__nav js-toggle" data-toggle="js-toggle__btn">
+            <a class="eg__nav-item" href="#">Home</a>
+            <a class="eg__nav-item" href="#">About</a>
+            <a class="eg__nav-item" href="#">Contact</a>
+            <button class="js-toggle__btn eg__nav-close" aria-label="Toggle off-canvas navigation">
+                <svg fill="#191919" height="24" viewBox="0 0 24 24" width="24" xmlns="http://www.w3.org/2000/svg">
+                    <path
+                        d="M19 6.41L17.59 5 12 10.59 6.41 5 5 6.41 10.59 12 5 17.59 6.41 19 12 13.41 17.59 19 19 17.59 13.41 12z" />
+                    <path d="M0 0h24v24H0z" fill="none" />
+                </svg>
+            </button>
+        </nav>
+    </main>
+</body>
+
 </html>
