+++ packages/toggle/tools/playwright.webpack.config.js
@@ -1,0 +1,8 @@
+const baseConfig = require('./webpack.config');
+
+module.exports = {
+    ...baseConfig,
+    devServer: {
+        port: 8089
+    }
+};
